+++ PetShop_Upgrade/DTOS/Products/Admin/AdminProductItemDTO.cs
@@ -6,8 +6,8 @@ namespace PetShop_Upgrade.DTOS.Products.Admin
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public double SellingPrice { get; set; }
-        public double ImportPrice { get; set; }
+        public decimal SellingPrice { get; set; }
+        public decimal ImportPrice { get; set; }
         public ProductType Type { get; set; }
         public IsActive IsActive { get; set; }
         public string CategoryName { get; set; }
